requirements.txt
openai-whisper==20231117
python-dotenv==1.0.0
tqdm==4.66.1
pytz==2024.1
requests==2.31.0
torch==2.1.0
numpy==1.24.3
pydub==0.25.1 